LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/url-resolution-expected.txt


PASS Unregistered property resolves against document (URL token) 
PASS Unregistered property resolves against document (URL function) 
FAIL Registered non-inherited <url> resolves against sheet (URL token) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered non-inherited <url> resolves against sheet (URL function) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered inherited <url> resolves against sheet (URL token) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered inherited <url> resolves against sheet (URL function) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered inherited <url> resolves against sheet (Child node, URL token) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered inherited <url> resolves against sheet (Child node, URL function) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered property with unregistered var reference resolves against sheet (URL token) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered property with unregistered var reference resolves against sheet. (URL function) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered property with registered var reference resolves against sheet of referenced property (URL token) assert_equals: expected "/css/css-properties-values-api/support/alt" but got "/css/css-properties-values-api"
FAIL Registered property with registered var reference resolves against sheet of referenced property (URL function) assert_equals: expected "/css/css-properties-values-api/support/alt" but got "/css/css-properties-values-api"
FAIL Unregistered property with registered var reference resolves against sheet of referenced property (URL token) assert_equals: expected "/css/css-properties-values-api/support/alt" but got "/css/css-properties-values-api"
FAIL Unregistered property with registered var reference resolves against sheet of referenced property (URL function) assert_equals: expected "/css/css-properties-values-api/support/alt" but got "/css/css-properties-values-api"
FAIL Multiple (registered) var reference resolve against respective sheets (URL token) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Multiple (registered) var reference resolve against respective sheets (URL function) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered UTF16BE-encoded var reference resolve against sheet (URL token) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
FAIL Registered UTF16BE-encoded var reference resolve against sheet (URL function) assert_equals: expected "/css/css-properties-values-api/support/main" but got "/css/css-properties-values-api"
